Moved the qf config to qf.vim
This setting that I was trying out, we moved it to qf.vim
This problem occured with me when I was using quickfix window with
Tagbar. Sometimes my qf window will appear beneath the small tagbar
window in the side and that would be very inconvienient.
I don't understand completly what this does, but it gets the job done
for now.