Fetch title for a link and paste it in markdown
I find myself pasting a link with it's title multiple times when I am
writing markdown file for my website or for my notes.
This is a handy mapping which fetches the title with a curl command and
format it to correct markdown format.
So a link paste is just <leader>lp away.