Updated snip functions to take commit title
So I wanted to use a cousin of snip called `dsa` to take notes for dsa.
All the commits I do for that, I wanted it to have (dsa) in the commit
message instead of (wiki) or (draft). So now the parameter is passed as
an argument to the relevant funcions.