Polyglot settings to sensible
now I don't remember exactly why I did this, but I think it was because
polyglot was changing either `softtabstop`, `shiftwidth` or `tabstop` to
`8` spaces in some c files and it was looking ugly.
So we switched of what polyglot thought sensible and just enabled the
configs we wanted in our own vimrc file. The main one is `autoread`.

This reads the file back into vim buffer when it is modified out of the
vim. This happens fairly frequently in my workflow, for example `make
format` will format the file in the terminal.