(vim) search man pages within vim
Now this is very interesting. vim by default supports :Man command from
ftplugin/man.vim
We change the behaviour of `K` to open the man page in vert buf within
vim
And since we have fzf, let's use fzf to search man pages. This is not
supremely refined but it works quite well