(vim) terminal scrollback should be long
When you run something in the terminal via :terminal command, the
scrollback of the output is by default 10k lines. After the limit is
reached, vim truncates the top 10% of the output to make the space for
new output. Read more in `:help twsl`.
We increase that scrollback to 100k.